+++ src/OGAMMAIN.cpp
@@ -21,6 +21,7 @@
 //Filename    : OGAMMAIN.CPP
 //Description : Main Game Object - Main menu
 
+#include "ambition/Ambition_config.hh"
 #include "ambition/Ambition_version.hh"
 #include "ambition/Ambition_vga.hh"
 
@@ -217,6 +218,8 @@ void Game::main_menu()
 			pointingOption = newPointingOption;
 		}
 		// ######### end Gilbert 23/7 ##########//
+
+      Ambition::drawModeInformation();
 
 		sys.blt_virtual_buf();		// blt the virtual front buffer to the screen